// frontend/src/app/enums/account-path.enum.ts
export enum AccountPathEnum {
  AUTH_PATH = 'auth',
  LOGIN_PATH = 'login',
  REGISTER_PATH = 'register',
  ACCOUNT = 'account',
  REFRESH = 'refresh',
  MY_ACCOUNT_PATH = 'my-account',
  RESET = 'reset',
  CHANGE_PASSWORD = 'change-password',
  VALIDATE = 'validate',
  SEND_RESET_PASSWORD_EMAIL = 'send-reset-password-email',
  LOGOUT = 'logout',
}